moving office_to_frames and to_frames to image_converter
